requirements.txt
python-telegram-bot==20.7
google-genai==0.1.0
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
pdfplumber==0.10.2
numpy==1.26.0
